src/fx-agent/requirements.txt
azure-ai-projects>=1.0.0b4
azure-identity>=1.19.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
